+++ sdk/schemaregistry/azure-schemaregistry/CHANGELOG.md
@@ -6,12 +6,13 @@ This version and all future versions will require Python 3.7+, Python 3.6 is no 
 
 ### Features Added
 
-- `get_schema_by_version` method has been added to the sync and async `SchemaRegistryClient`.
+- `group_name`, `name`, and `version` have been added as optional parameters to the method `get_schema` method on the sync and async `SchemaRegistryClient`.
 - `version` has been added to `SchemaProperties`.
 
 ### Other Changes
 
 - Updated azure-core minimum dependency to 1.24.0.
+- Added distributed tracing support for sync and async `SchemaRegistryClient`.
 
 ## 1.1.0 (2022-05-10)